fix(root): rename heartly-cli, eslint updates
