+++ package.json
@@ -46,16 +46,26 @@
     "@commitlint/cli": "^11.0.0",
     "@commitlint/config-lerna-scopes": "^10.0.0",
     "@commitlint/prompt": "^11.0.0",
-    "@heartly/cli": "file:packages/heartly",
-    "@heartly/eslint-config-typescript-react": "file:packages/eslint-config-typescript",
-    "@heartly/prettier-config": "file:packages/prettier-config",
     "@types/node": "^14.0.0",
+    "@typescript-eslint/eslint-plugin": "^4.0.0",
+    "@typescript-eslint/parser": "^3.1.0",
     "commitizen": "^4.0.3",
     "concurrently": "^5.3.0",
     "conventional-changelog": "^3.1.18",
     "eslint": "^7.0.0",
+    "eslint-config-standard": "^14.1.0",
+    "eslint-plugin-import": "^2.20.1",
+    "eslint-config-prettier": "^6.2.0",
+    "eslint-plugin-filenames": "^1.3.2",
+    "eslint-plugin-jest": "^24.0.0",
+    "eslint-plugin-json": "^2.1.0",
+    "eslint-plugin-node": "^11.0.0",
+    "eslint-plugin-prettier": "^3.1.0",
+    "eslint-plugin-promise": "^4.2.1",
+    "eslint-plugin-react": "^7.19.0",
+    "eslint-plugin-standard": "^4.0.1",
     "husky": "^4.0.2",
-    "lerna": "^3.10.7",
+    "lerna": "^3.20.2",
     "lint-staged": "^10.0.8",
     "markdown-spellcheck": "^1.3.1",
     "markdownlint": "^0.21.0",
