+++ packages/heartly-cli/README.md
@@ -1,0 +1,32 @@
+# @heartly/heartly-cli 💕
+
+> This readme is automatically generated by [create-ink-app](https://github.com/vadimdemedes/create-ink-app)
+
+
+## Install
+
+```sh
+yarn add --global @heartly/heartly-cli
+```
+
+
+## CLI
+
+```
+$ cli --help
+
+  Usage
+    $ cli
+
+  Options
+    --name  Your name
+
+  Examples
+    $ cli --name=Jane
+    Hello, Jane
+```
+
+## Development
+
+For most development needs, view the great [link docs](https://github.com/vadimdemedes/ink). 
+To make the CLI locally executable, run `yarn link` within this folder.
