+++ packages/heartly-cli/src/index.tsx
@@ -1,0 +1,28 @@
+#!/usr/bin/env node
+import React from 'react'
+import { render } from 'ink'
+import meow from 'meow'
+import App from './ui'
+
+const cli = meow(
+  `
+	Usage
+	  $ cli
+
+	Options
+		--name  Your name
+
+	Examples
+	  $ cli --name=Jane
+	  Hello, Jane
+`,
+  {
+    flags: {
+      name: {
+        type: 'string',
+      },
+    },
+  },
+)
+
+render(<App name={cli.flags.name} />)
