--- packages/heartly/README.md
@@ -1,27 +1,0 @@
-# cli
-
-> This readme is automatically generated by [create-ink-app](https://github.com/vadimdemedes/create-ink-app)
-
-
-## Install
-
-```bash
-$ npm install --global cli
-```
-
-
-## CLI
-
-```
-$ cli --help
-
-  Usage
-    $ cli
-
-  Options
-    --name  Your name
-
-  Examples
-    $ cli --name=Jane
-    Hello, Jane
-```
