--- packages/heartly/tsconfig.json
@@ -1,11 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "compilerOptions": {
-    "declaration": false,
-    "esModuleInterop": true,
-    "outDir": "dist",
-    "sourceMap": false
-  },
-  "include": ["src"],
-  "exclude": ["**/*.test.tsx"]
-}
